feat(test): criando testes para os domain services

+++ internal/domain/quote/quote.go
@@ -155,7 +155,7 @@ type Offer struct {
 
 type Metrics struct {
 	Carrier         string
-	TotalOffers     int
+	TotalOffers     float64
 	AvgPrice        float64
 	IsCheapest      bool
 	IsMostExpensive bool
